actualization

+++ org.adempiere.base/src/org/compiere/dbPort/Convert.java
@@ -37,7 +37,6 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import org.compiere.db.Database;
-import org.compiere.model.MSysConfig;
 import org.compiere.util.CLogger;
 import org.compiere.util.DisplayType;
 import org.compiere.util.Env;
@@ -306,7 +305,7 @@ public abstract class Convert
 		Pattern p = Pattern.compile("'[[^']*]*'");
 		Matcher m = p.matcher(inputValue);
 		int i = 0;
-		StringBuffer retValue = new StringBuffer(inputValue.length());
+		StringBuilder retValue = new StringBuilder(inputValue.length());
 		while (m.find()) {
 			String var = inputValue.substring(m.start(), m.end()).replace(quoteMarker, "''"); // Put back quotes, if any
 			retVars.addElement(var);
@@ -438,13 +437,7 @@ public abstract class Convert
 	public synchronized static void logMigrationScript(String oraStatement, String pgStatement) {
 		// Check AdempiereSys
 		// check property Log migration script
-		boolean logMigrationScript = false;
-		if (Ini.isClient()) {
-			logMigrationScript = Ini.isPropertyBool(Ini.P_LOGMIGRATIONSCRIPT);
-		} else {
-			String sysProperty = Env.getCtx().getProperty("LogMigrationScript", "N");
-			logMigrationScript = "y".equalsIgnoreCase(sysProperty) || "true".equalsIgnoreCase(sysProperty);
-		}
+		boolean logMigrationScript = isLogMigrationScript();
 		if (logMigrationScript) {
 			if (dontLog(oraStatement))
 				return;
@@ -480,6 +473,20 @@ public abstract class Convert
 		}
 	}
 
+	/**
+	 * @return true if it is in log migration script mode
+	 */
+	public static boolean isLogMigrationScript() {
+		boolean logMigrationScript = false;
+		if (Ini.isClient()) {
+			logMigrationScript = Ini.isPropertyBool(Ini.P_LOGMIGRATIONSCRIPT);
+		} else {
+			String sysProperty = Env.getCtx().getProperty("LogMigrationScript", "N");
+			logMigrationScript = "y".equalsIgnoreCase(sysProperty) || "true".equalsIgnoreCase(sysProperty);
+		}
+		return logMigrationScript;
+	}
+
 
 	private static String [] dontLogTables = new String[] {
 			"AD_ACCESSLOG",
@@ -507,6 +514,7 @@ public abstract class Convert
 			"AD_WINDOW_ACCESS",
 			"AD_WORKFLOW_ACCESS",
 			"AD_WORKFLOWPROCESSORLOG",
+			"AD_USERPREFERENCE",
 			"CM_WEBACCESSLOG",
 			"C_ACCTPROCESSORLOG",
 			"K_INDEXLOG",
@@ -541,6 +549,9 @@ public abstract class Convert
 			return true;
 		if (uppStmt.startsWith("UPDATE R_REQUESTPROCESSOR SET DATELASTRUN"))
 			return true;
+		// don't log sequence updates
+		if (uppStmt.startsWith("UPDATE AD_SEQUENCE SET CURRENTNEXT"))
+			return true;
 		// Don't log DELETE FROM Some_Table WHERE AD_Table_ID=? AND Record_ID=?
 		if (uppStmt.startsWith("DELETE FROM ") && uppStmt.endsWith(" WHERE AD_TABLE_ID=? AND RECORD_ID=?"))
 			return true;
@@ -573,13 +584,8 @@ public abstract class Convert
 
 	private static void writeLogMigrationScript(Writer w, String statement) throws IOException
 	{
-		boolean isUseCentralizedID = "Y".equals(MSysConfig.getValue(MSysConfig.DICTIONARY_ID_USE_CENTRALIZED_ID, "Y")); // defaults to Y
-		boolean isUseProjectCentralizedID = "Y".equals(MSysConfig.getValue(MSysConfig.PROJECT_ID_USE_CENTRALIZED_ID, "N")); // defaults to N
 		String prm_COMMENT;
-		if (!isUseCentralizedID && isUseProjectCentralizedID)
-			prm_COMMENT = MSysConfig.getValue(MSysConfig.PROJECT_ID_COMMENTS);
-		else
-			prm_COMMENT = MSysConfig.getValue(MSysConfig.DICTIONARY_ID_COMMENTS);
+		prm_COMMENT = Env.getContext(Env.getCtx(), "MigrationScriptComment");
 		if (prm_COMMENT != null && ! m_oldprm_COMMENT.equals(prm_COMMENT)) {
 			// log sysconfig comment
 			w.append("-- ");
